working on new rebalancer
