+++ src/config.ts
@@ -22,6 +22,7 @@ export const PERIOD_LENGTH: number = 5; // in Ethereum blocks
 export const PERIOD_LIMIT: number = 75000; // transactions allowed per period
 export const STAKE_CONTRACT_ADDR: string = "0x67b2821031489750080cE33C86ba6Ea97eEC7763";
 export const STAKE_CONTRACT_ABI: Array<object> = [{"constant":true,"inputs":[{"name":"a","type":"address"}],"name":"stakeFor","outputs":[{"name":"","type":"uint256"}],"payable":false,"stateMutability":"view","type":"function"},{"constant":true,"inputs":[],"name":"totalStaked","outputs":[{"name":"","type":"uint256"}],"payable":false,"stateMutability":"view","type":"function"},{"constant":false,"inputs":[{"name":"amount","type":"uint256"}],"name":"removeStake","outputs":[],"payable":false,"stateMutability":"nonpayable","type":"function"},{"constant":false,"inputs":[{"name":"amount","type":"uint256"}],"name":"stake","outputs":[],"payable":false,"stateMutability":"nonpayable","type":"function"},{"constant":true,"inputs":[],"name":"digm","outputs":[{"name":"","type":"address"}],"payable":false,"stateMutability":"view","type":"function"},{"inputs":[{"name":"_digm","type":"address"}],"payable":false,"stateMutability":"nonpayable","type":"constructor"},{"anonymous":false,"inputs":[{"indexed":false,"name":"staker","type":"address"},{"indexed":false,"name":"amount","type":"uint256"}],"name":"StakeMade","type":"event"},{"anonymous":false,"inputs":[{"indexed":false,"name":"staker","type":"address"},{"indexed":false,"name":"amount","type":"uint256"}],"name":"StakeRemoved","type":"event"}];
+export const FINALITY_THRESHOLD: number = 2;
 
 // endpoints for ABCI/RPC
 export const ABCI_HOST: string = "localhost";
