+++ src/other/index.old_10-27.ts
@@ -5,57 +5,38 @@
   =========================
 
   @date_inital 12 September 2018
-  @date_modified 27 October 2018
+  @date_modified 19 October 2018
   @author Henry Harder
 
   Entry point and startup script for ParadigmCore. 
-*/
 
-// Standard lib and 3rd party NPM modules
+
 import * as _ws from "ws";
 import * as tendermint from "tendermint-node";
-import { EventEmitter } from "events";
+import * as fs from "fs";
 
-// ParadigmCore classes
 import { Logger } from "./util/Logger";
 import { WebSocketMessage } from "./net/WebSocketMessage";
 import { messages as msg } from "./util/messages";
+import { EventEmitter } from "events";
 
-// State object templates
-import { deliverState as dState } from "./state/deliverState";
-import { commitState as cState} from "./state/commitState";
+import { deliverState } from "./state/deliverState";
+import { commitState } from "./state/commitState";
 
-// Initialization functions
-import { startMain, startRebalancer } from "./abci/main";
+import { startMain, startRebalancer } from "./abci/handlers";
 import { startAPIserver } from "./net/server";
 
-// Configuration and constants
-import { 
-    WS_PORT,
-    TM_HOME, 
-    ABCI_HOST, 
-    ABCI_RPC_PORT, 
-    API_PORT, 
-    WEB3_PROVIDER, 
-    PERIOD_LENGTH, 
-    PERIOD_LIMIT, 
-    STAKE_CONTRACT_ADDR, 
-    STAKE_CONTRACT_ABI, 
-    ABCI_PORT,
-    VERSION,
-    FINALITY_THRESHOLD
-} from "./config";
+import { WS_PORT, TM_HOME, ABCI_HOST, ABCI_RPC_PORT, API_PORT } from "./config";
 
-let wss: _ws.Server;        // OrderStream WS server
-let emitter: EventEmitter;  // Emitter to track events
-let node: any;              // Tendermint node instance
+
+let wss: _ws.Server;
+let emitter: EventEmitter;
+let node: any; // Tendermint node instance
 
 /**
- * This function executes immediately upon this file being loaded. It is
+ * This function executes immediately upon this file being executed. It is
  * responsible for starting all dependant modules.
- * 
- * Provide configuration options via `config.ts`
- */
+ *
 (async function() {
     Logger.logStart();
 
@@ -68,9 +49,13 @@ let node: any;              // Tendermint node instance
                 laddr: `tcp://${ABCI_HOST}:${ABCI_RPC_PORT}`
             }
         });
+
+        // node.stdout.pipe(process.stdout); // pipe tendermint logs to STDOUT
+
     } catch (error) {
-        Logger.consensusErr(msg.abci.errors.tmFatal);
-        process.exit(1);
+        console.log(error);
+        Logger.consensusErr("Fatal error starting Tendermint core.");
+        process.exit();
     }
 
     // Start WebSocket server
@@ -81,45 +66,25 @@ let node: any;              // Tendermint node instance
         });
         emitter = new EventEmitter(); // parent event emitter
     } catch (error) {
-        Logger.websocketErr(msg.websocket.errors.fatal);
-        process.exit(1);
+        Logger.websocketErr("Fatal error starting WebSocket server.");
+        process.exit();
     }
 
     // Start ABCI application
     try{
-        let options = {
-            // ABCI configuration options
-            "emitter": emitter,
-            "deliverState": dState,
-            "commitState": cState,
-            "version": VERSION,
-            "abciServPort": ABCI_PORT,
-        
-            // Rebalancer options
-            "provider": WEB3_PROVIDER,
-            "periodLength": PERIOD_LENGTH,
-            "periodLimit": PERIOD_LIMIT,
-            "finalityThreshold": FINALITY_THRESHOLD,
-            "stakeAddress": STAKE_CONTRACT_ADDR,
-            "stakeABI": STAKE_CONTRACT_ABI,
-            "abciHost": ABCI_HOST,
-            "abciPort": ABCI_RPC_PORT
-        }
-
-        // Wait for main process to start
-        await startMain(options);
+        // start main ParadigmCore logic 
+        await startMain(deliverState, commitState, emitter);
         Logger.consensus("Waiting for Tendermint to synchronize...");
 
-        // Wait for Tendermint to load and synchronize
         await node.synced();
         Logger.consensus("Tendermint initialized and syncronized.");
 
-        // Start state rebalancer sub-process AFTER sync
+        // start state rebalancer sub-process AFTER sync
         await startRebalancer();
-        Logger.rebalancer(msg.rebalancer.messages.activated, 0);
+        Logger.rebalancer("Stake rebalancer activated. Subscribed to Ethereum events.", 0);
     } catch (error) {
         Logger.logError(msg.abci.errors.fatal);
-        process.exit(1);
+        process.exit();
     }
 
     // Start HTTP API server
@@ -128,14 +93,14 @@ let node: any;              // Tendermint node instance
         await startAPIserver(ABCI_HOST, ABCI_RPC_PORT, API_PORT);
     } catch (error) {
         Logger.apiErr(msg.api.errors.fatal)
-        process.exit(1);
+        process.exit();
     }
 
     /**
      * Begin WebSocket handler implementation (below)
      * 
      * TODO: move this to another file
-     */
+     *
 
     wss.on("connection", ws => {
         try {
@@ -155,18 +120,6 @@ let node: any;              // Tendermint node instance
                 Logger.websocketErr(msg.websocket.errors.broadcast);
             }
         });
-
-        /*emitter.on("stream", stream => {
-            try {
-                wss.clients.forEach(client => {
-                    if ((client.readyState === 1) && (client === ws)){
-                        WebSocketMessage.sendStream(client, stream);
-                    }
-                });
-            } catch (err) {
-                Logger.websocketErr(msg.websocket.errors.broadcast);
-            }
-        });*/
         
         ws.on('message', message => {
             if(message === "close") { 
@@ -177,5 +130,5 @@ let node: any;              // Tendermint node instance
         });
     });
 
-    Logger.logEvent(msg.general.messages.start);
-})();
+    Logger.logEvent("Initialization complete, begining new block production.");
+})();*/
