+++ src/state/deliverState.ts
@@ -5,7 +5,7 @@
   =========================
   
   @date_inital 25 September 2018
-  @date_modified 19 October 2018
+  @date_modified 24 October 2018
   @author Henry Harder
 
   Object that represents the initial and pre-commit state of the OS node.
@@ -15,9 +15,13 @@ export let deliverState = {
   round: {
     number: 0,
     startsAt: 0,
-    endsAt: 0
+    endsAt: 0,
+    limit: 0
   },
-  mapping: {},
+  events: {},
+  balances: {},
+  limits: {},
   orderCounter: 0,
-  lastBlockHeight: 0
+  lastBlockHeight: 0,
+  lastBlockAppHash: null
 }
