+++ src/util/Codes.ts
@@ -1,0 +1,11 @@
+export default class Codes {
+    static OK           = 0;
+    static NO_BLOCK     = 1;
+    static CONTRACT     = 2;
+    static ABCI_CON     = 3;
+    static URI_SCHEME   = 4;
+    static WEB3_PROV    = 5;
+    static WEB3_INST    = 6;
+    static SUBSCRIBE    = 7;
+    static NO_ABCI      = 8;
+}
