+++ src/util/Timestamp.ts
@@ -5,7 +5,7 @@
   =========================
   
   @date_inital 10 October 2018
-  @date_modified 19 October 2018
+  @date_modified 27 October 2018
   @author Henry Harder
 
   Dumb class for creating a log timestamp.
@@ -17,7 +17,6 @@ export class Timestamp extends Date {
     }
 
     public logFormat() {
-        
         return `${Math.floor(Date.now()/1000).toString()}.${Date.now().toString().slice(-3)}`;
     }
 }
